+++ crates/ruma-client-api/CHANGELOG.md
@@ -5,7 +5,9 @@ Breaking changes:
 * Remove `PartialEq` implementations for a number of types
   * If the lack of such an `impl` causes problems, please open a GitHub issue
 * Split `uiaa::UserIdentifier::ThirdParty` into two separate variants
-* Make `message::get_message_events::v3::Request::new`'s `from` parameter optional
+* Remove `message::get_message_events::v3::Request::new`'s `from` parameter
+  * This parameter is now optional, and can be set after initial construction, or by using one of
+    the other constructors
 * `receipt::create_receipt` uses its own `ReceiptType`
 
 Improvements: